requirements.txt
# Data pipeline - Olist E-Commerce (PostgreSQL)
# Python 3.10+

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# Ingestion & ETL
pandas>=2.0.0
openpyxl>=3.1.0

# Data quality
great-expectations>=0.18.0

# Analysis
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Orchestration (optional)
prefect>=2.14.0

# Environment (optional)
python-dotenv>=1.0.0
